requirements.txt
asgiref==3.2.7
Django==3.0
django-factory==0.11
djangorestframework==3.11.0
extras==1.0.0
factory-boy==2.12.0
Faker==4.1.0
fixtures==3.0.0
linecache2==1.0.0
pbr==5.4.5
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
testtools==2.4.0
text-unidecode==1.3
traceback2==1.4.0
unittest2==1.1.0
